pyproject.toml
[tool.poetry]
name = "dsl"
version = "1.0.0"
description = "An implementation of Semantic labeling: A domain-independent approach"
authors = ["Binh Vu <binh@toan2.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/binh-vu/dsl"

[tool.poetry.dependencies]
python = "^3.11"
sem-desc = "^5.5.0"
fastnumbers = "^5.0.1"
ftfy = "^6.1.3"
spacy = "^3.7.2"
tqdm = "^4.66.1"
serde2 = "^1.6.0"
scipy = "^1.11.4"
numpy = "^1.26.2"
scikit-learn = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
